@@ -48,12 +48,18 @@
 
 FactoryGirl.define do
   factory :customers_auth, class: CustomersAuth do
-    customer_id nil
-    rateplan_id nil
-    enabled true
-    ip nil
-    account_id nil
-    gateway_id nil
+    sequence(:name) { |n| "customers_auth_#{n}"}
+    ip '0.0.0.0'
+    diversion_policy_id 1
+    dump_level_id 1
+
+    association :customer, factory: :contractor, customer: true
+    association :rateplan
+    association :routing_plan
+    association :gateway
+    association :account
+
+
     src_rewrite_rule nil
     src_rewrite_result nil
     dst_rewrite_rule nil
@@ -61,20 +67,14 @@ FactoryGirl.define do
     src_prefix ''
     dst_prefix ''
     x_yeti_auth nil
-    name nil
-    dump_level_id 1
-    capacity 0
     pop_id nil
     uri_domain nil
     src_name_rewrite_rule nil
     src_name_rewrite_result nil
-    diversion_policy_id 1
     diversion_rewrite_rule nil
     diversion_rewrite_result nil
-    dst_blacklist_id nil
-    src_blacklist_id nil
-    routing_group_id nil
-    routing_plan_id nil
+    dst_numberlist_id nil
+    src_numberlist_id nil
     allow_receive_rate_limit false
     send_billing_information false
   end
